feat(arguments): add command argument for better command parsing
Prior to this, commands could only be fetched if they were already loaded. Now there is an argument
that parses them automatically, including a glob that will search for the command file and fetch the
command if it isn't already loaded. On top of this, the commit stylizes a missing semicolon in the
ping command, and adds a new dependency (`globby`).