feat(commands): add classes that implement categories and helpers
This commit is a major breaking change. It implements helper functions for command categories, as
well as the categorization system itself, by adding two new classes for a modified command store and
command class. As a result, all files had to be changed to use the new command class too. All
functions have been tested to work as of 23/10/2020.

BREAKING CHANGE: Two new classes have been implemented into the client and changed project-wide for
a new command store and commands. This is not backwards compatible and modules written from this
branch forwards will not work on older versions of Syrus that do not have these new classes.

RE #9 (1, 2, and 3)