feat(music): implement music manager
This implements a music manager (lavaqueue) with a Redis-based queues framework. It is initialized
when `events/onceReady` triggers because it needs the client ID which isn't available until login is
complete. The music manager has, however, been stripped of all its old features, and is no longer an
extension of guilds (it is HIGHLY inefficient to create one instance of the manager per guild, so we
are now resorting to a global music framework which can operate on guilds individually given their
ID).

RE #13