feat(music): add fetchTracks and fix websocket handler
This commit adds the `fetchTracks` function (parameters: `query`), which automatically checks if the query is a URL and if not sets it up to be a Yo
uTube search query; then, it returns the result of passing the modified query to Lavalink. Additionally, this commit fixes the websocket handler, rem
oving the guild cache checker and replacing it with solely raw websocket downstream events. Furthermore, SSL is implemented into Lavalink configurati
on and also the Lavalink websocket handler is added.

RE #13