fix(commands): prevent play from starting queue if already started
This commit adds a check to the play command to wrap the part where it begins the queue. Before, it
would call indiscriminately, which causes a break in the audio when adding new songs to the queue if
one is already playing. This is considered unexpected behaviour, and so now to guard unnecessary
calls and prevent this break, a check is added so it only starts the queue if it is not already
started.

RE #13