latest staging -> master (#28)
* feat(core): add guildDelete to database driver

* feat(core): add guildDelete event

This implements the `guildDelete` event, which removes guilds from  the database when the bot is
removed from them. This is to preserve performance and space for the database.

RE #1

* refactor(commands): add formatting to set

RE #1

* fix(commands): fix argument parsing in set

This fixes the error produced by the set command when specifying a key with no value, as it would
only check for the presence of a key before continuing.

* fix: address LGTM concerns (#27)


addresses all LGTM concerns for the bot.