push core changes upstream (#35)
* fix(arguments): address LGTM concerns

* fix(cmd:core): address LGTM concerns

* fix(cmd:fun): address LGTM concerns

* fix(cmd:moderation): address LGTM concerns

* fix(cmd:music): address LGTM concerns

* fix(cmd:utility): address LGTM concerns

* fix(core): address LGTM concerns

* fix(predicates): address LGTM concerns

* fix(db): address LGTM concerns

* refactor(pm): update CONTRIBUTING.md

* perf(cmd:utility): remove unused vars in info

* feat(core): add elegant exit events

This implements exit events that enable components of the bot to clean up gracefully by listening to
the `cleanup` event when the process exits. These are triggered by any exit, and special cases are
made for uncaught exceptions or SIGINT.

RE #34

* ci(eslint): enable process.env

* feat(core): add cutoff safeguard for production mode

Prior to this, production mode's uncaught exceptions would follow the same logic as development.
This introduces a conditional that implements 'the show must go on' logic - uncaught exceptions will
print to console, but will NOT trigger an exit when in production mode.

* feat(music): add connection failure handling

RE #34

* feat(db): add cleanup sequence

* fix(music): add support for resolution failure

Prior to this, the connection error handler would only check for the ECONNREFUSED event. This is
fine, assuming the address exists - however, when it does not, a different error is thrown, which is
checked for with this fix.

* fix(music): handle queue finish properly

Prior to this, the queue finish embed would fire after every song. This is because the STOPPED and
FINISHED events are not in relation to the queue, but rather one track. Now, it checks the length
after an end event.