feat(checks): add full suite cascading predicates (#42)
* feat(db): add roles boilerplate arrays to guild schema

This adds the `modRoles`, `adminRoles` and `djRoles` arrays (`Array<Number>`) to the guild schema of
the database driver.

RE #25

* feat(checks): add mod, admin and dj preconditions

RE #25

* feat(checks): add cascades for preconditions

RE #25

* revert(checks): remove custom permissions precondition

RE #25

* refactor(checks): use builtins for owner precondition

Prior to this, the Owner precondition would import UserError, ok, and err to handle its return
system. Now, it uses the precondition builtins.

* refactor(checks): reformat precondition names

* perf(checks): remove unused db call in serverowner

Prior to this, ServerOwner would instantiate a guild instance from Syrus' configuration. However, it
was not actually used or passed anywhere. This fixes that unnecessary call.

* fix(checks): asynchronise owner

* perf(checks): integrate configuration passing

Prior to this, each call up the cascade would instantiate a new configuration instance. Now, the
database will only be queried once, and from there the configuration will be passed along the chain.