test(db): fix global collection tests
This removes the unnecessary clunk from the global instantiation test that previously existed since
it would use the same pre-existing database as the one in the bot's main config. It also patches the
find existing global test as prior to this it would just instantiate through the driver rather than
making a new one and testing if the driver would find it.