perf(core): make command categories static
Prior to this, every time `commands.categories` was accessed, it would dynamically return an array
computed by iterating through all commands and logging their unique categories. Now, hooks are
implemented into the store, so the array is static and recomputed only when a command is loaded or
unloaded. This cuts access times approximately in half.