test: implement unit testing (#49)
* refactor: apply new lint rules

RE #41

* ci(lint): standardise rules

RE #41

* chore: update deps

This is in response to dependabot alerts LOW:ini and MOD:mquery.

RE #38

* chore(pm): add nyc_output to gitignore

* chore(package): update all npm dependencies (2020-12-21) (#47)

Co-authored-by: depfu[bot] <23717796+depfu[bot]@users.noreply.github.com>

* chore(deps): reapply security updates (#38)

* chore(deps): prepare to add unit testing

This adds the following development dependencies: `mocha` (unit testing framework), `nyc` (coverage
reporter), and `sinon` (function mocking and spying framework).

* feat(pm): set up unit testing

* test(preconditions): add mongodb driver

* ci(gh): modernise workflow actions and add new for code scanning & unit testing

* chore(deps): add mongodb-memory-server

* feat(db): add ability to provide connection string to driver

* fix(db): add in-memory mongo instance to testing

* test(db): fix global collection tests

This removes the unnecessary clunk from the global instantiation test that previously existed since
it would use the same pre-existing database as the one in the bot's main config. It also patches the
find existing global test as prior to this it would just instantiate through the driver rather than
making a new one and testing if the driver would find it.

* test(db): include global in cleanup to fix duplicates in global tests

* test(db): add stricter timings

* refactor(core): shift config dependency out of core modules

RE #48

* test(db): remove unnecessary schema from test file

* refactor(core): adjust default branching

This adds a default configuration for the music manager in preparation for unit testing and moves
the connection string default out of the database driver.

RE #48

* refactor(core): move default config

* test(db): allow extra spin-up time for in memory server

* refactor(core): move db init out of constructor into separate function

* test(db): remove fixed spin-up timeout

This would cause tests to fail when mongodb-memory-server would take too long to fire up. This is
undesirable, as mongodb-memory-server does not exist in a production environment, and is a variable
exclusive solely to our testing environment - therefore, its timing should not be factored into our
tests.

* test(core): add tests for command framework

* test(core): finalise command framework tests

* test: cover core & db, lint

* perf(core): optimize categorization of command store

* feat(cmd): add timing to eval

* perf(core): make command categories static

Prior to this, every time `commands.categories` was accessed, it would dynamically return an array
computed by iterating through all commands and logging their unique categories. Now, hooks are
implemented into the store, so the array is static and recomputed only when a command is loaded or
unloaded. This cuts access times approximately in half.

* test(core): add testing for new hook