refactor: migrate to latest sapphire
This migrates to the latest Sapphire version (1.0.0-alpha.2). A lot of breaking changes are made, so
buckle up: LoaderStrategy was added, with the onLoad and onUnload methods to implement static
command categories. Default i18n no longer exists, so @sapphire/plugin-i18next is used instead -
along with the Message#editTranslated, Message#replyTranslated, and Channel#sendTranslated methods.
Along the way, the info and serverinfo commands were fixed to increase performance. Another bug with
Music/disconnect, where the bot would error if the user was not in a channel when running the
command, was fixed too. All this, and now module-alias has been added for namespacing. Welcome to
the latest Syrus :)

BREAKING CHANGE: All of the above changes are considered breaking.