refactor: update to latest sapphire (#55)
* refactor: apply new lint rules

RE #41

* chore: update deps

This is in response to dependabot alerts LOW:ini and MOD:mquery.

RE #38

* chore(package): update all npm dependencies (2020-12-21) (#47)

Co-authored-by: depfu[bot] <23717796+depfu[bot]@users.noreply.github.com>

* chore(deps): prepare to add unit testing

This adds the following development dependencies: `mocha` (unit testing framework), `nyc` (coverage
reporter), and `sinon` (function mocking and spying framework).

* test(preconditions): add mongodb driver

* chore(deps): add mongodb-memory-server

* feat(db): add ability to provide connection string to driver

* fix(db): add in-memory mongo instance to testing

* test(db): fix global collection tests

This removes the unnecessary clunk from the global instantiation test that previously existed since
it would use the same pre-existing database as the one in the bot's main config. It also patches the
find existing global test as prior to this it would just instantiate through the driver rather than
making a new one and testing if the driver would find it.

* test(db): include global in cleanup to fix duplicates in global tests

* test(db): add stricter timings

* refactor(core): shift config dependency out of core modules

RE #48

* test(db): remove unnecessary schema from test file

* refactor(core): adjust default branching

This adds a default configuration for the music manager in preparation for unit testing and moves
the connection string default out of the database driver.

RE #48

* refactor(core): move default config

* revert(db): remove default config

* revert(music): remove default config

* test(db): allow extra spin-up time for in memory server

* refactor(core): move db init out of constructor into separate function

* test(db): remove fixed spin-up timeout

This would cause tests to fail when mongodb-memory-server would take too long to fire up. This is
undesirable, as mongodb-memory-server does not exist in a production environment, and is a variable
exclusive solely to our testing environment - therefore, its timing should not be factored into our
tests.

* test(core): add tests for command framework

* test(core): finalise command framework tests

* test: cover core & db, lint

* perf(core): optimize categorization of command store

* feat(cmd): add timing to eval

* perf(core): make command categories static

Prior to this, every time `commands.categories` was accessed, it would dynamically return an array
computed by iterating through all commands and logging their unique categories. Now, hooks are
implemented into the store, so the array is static and recomputed only when a command is loaded or
unloaded. This cuts access times approximately in half.

* test(core): add testing for new hooks

* chore(deps): update sapphire

* refactor(i18n): reformat language keys

* refactor(core): use LoaderStrategy for command framework

* test(cmd): fix class testing

* refactor: migrate to latest sapphire

This migrates to the latest Sapphire version (1.0.0-alpha.2). A lot of breaking changes are made, so
buckle up: LoaderStrategy was added, with the onLoad and onUnload methods to implement static
command categories. Default i18n no longer exists, so @sapphire/plugin-i18next is used instead -
along with the Message#editTranslated, Message#replyTranslated, and Channel#sendTranslated methods.
Along the way, the info and serverinfo commands were fixed to increase performance. Another bug with
Music/disconnect, where the bot would error if the user was not in a channel when running the
command, was fixed too. All this, and now module-alias has been added for namespacing. Welcome to
the latest Syrus :)

BREAKING CHANGE: All of the above changes are considered breaking.

* chore: fix dependency security

* chore: update dependencies

* chore(pm): add all-contributors spec

* docs: update @Nytelife26 as a contributor

* docs: update @mynameismax as a contributor

* docs: add @colecci as a contributor

* fix(checks): naming conventions & error handling

* refactor(events): reformat all

* chore(pm): fix readme license link

* test(data): remove accidental client inclusion

* feat(i18n): add language key for new event

* style(events): reformat event class names

* feat(events): add VoiceStateUpdate

This new event implements functionality to leave the voice channel once there are no members
remaining, to conserve bandwidth and processing power.

* refactor: utilize new store context

* ci(release): change release trigger branch to staging

Following this, staging will be the branch that triggers releases. Master will then be the branch
that hosts major stable versions, and the version that the docker image for the production-ready bot
is based on, meanwhile staging will instantiate Syrus Development.

* fix(events): address lgtm concerns

* ci: streamline release workflow

* style: fix typo in workflow